requirements.txt
face-recognition==1.3.0
opencv-python==4.8.0.76
requests==2.31.0
numpy==1.24.3
python-dotenv==1.2.1